# Performance tweaks on leveldb code and blockmap