This improves benchmarks and then makes a few small tweaks to the blockmap code to speed it up. The total difference is below, each commit contains it's benchmark difference against the previous one. The ReplaceAll benchmark represents a large Replace or Update and is sped up mostly by reducing allocations for keys in the blockmap code and doing batching. The UpdateOne benchmarks represents small update calls and are mostly sped up by removing the forced GC we did there. I don't think we need to do any forced GC calls.

 ```
benchmark                         old ns/op      new ns/op      delta
BenchmarkReplaceAll-8             2942370526     1868198122     -36.51%
BenchmarkUpdateOneChanged-8       7402489        231852         -96.87%
BenchmarkUpdateOneUnchanged-8     7298777        230624         -96.84%
BenchmarkNeedHalf-8               113608416      104601744      -7.93%
BenchmarkHave-8                   29834263       29102480       -2.45%
BenchmarkGlobal-8                 162773699      150547687      -7.51%
BenchmarkNeedHalfTruncated-8      111943400      102471355      -8.46%
BenchmarkHaveTruncated-8          29490369       28758368       -2.48%
BenchmarkGlobalTruncated-8        165841081      151192913      -8.83%

benchmark                         old allocs     new allocs     delta
BenchmarkReplaceAll-8             1054942        555577         -47.34%
BenchmarkUpdateOneChanged-8       1149           1135           -1.22%
BenchmarkUpdateOneUnchanged-8     1135           1135           +0.00%
BenchmarkNeedHalf-8               374774         374780         +0.00%
BenchmarkHave-8                   151995         151992         -0.00%
BenchmarkGlobal-8                 530042         530033         -0.00%
BenchmarkNeedHalfTruncated-8      374697         374699         +0.00%
BenchmarkHaveTruncated-8          151834         151834         +0.00%
BenchmarkGlobalTruncated-8        530050         530037         -0.00%

benchmark                         old bytes      new bytes      delta
BenchmarkReplaceAll-8             5074294728     1765116216     -65.21%
BenchmarkUpdateOneChanged-8       141048         135085         -4.23%
BenchmarkUpdateOneUnchanged-8     134976         134976         +0.00%
BenchmarkNeedHalf-8               44734813       44758752       +0.05%
BenchmarkHave-8                   11911634       11845052       -0.56%
BenchmarkGlobal-8                 80436854       80431136       -0.01%
BenchmarkNeedHalfTruncated-8      46514673       46526459       +0.03%
BenchmarkHaveTruncated-8          11348357       11348357       +0.00%
BenchmarkGlobalTruncated-8        81730740       80977672       -0.92%
```