@AudriusButkevicius 2015-10-21
Bravo. I'd like to see the final before and after benchmark just to feel the satisfaction :D